refactor: extract Pre-approved handler from handleStateTransitions
Move the Pre-approved branch into effects/preApprovedEffects.ts
(~105 lines). Handler pre-saves firstApprovedAt / firstApprovedBy
and the persisted XState snapshot to Firestore before updating the
Google Calendar event title with the [PRE-APPROVED] prefix.

No behavior change. tsc clean, 44 tests pass.
